move circuit analysis functions to extra file
